#190: 提案 done

+++ voter_guide/bills/views.py
@@ -10,6 +10,8 @@ from councilors.models import CouncilorsDetail
 from search.models import Keyword
 from search.views import keyword_list, keyword_been_searched
 from .models import Bills, Councilors_Bills
+from candidates.models import Intent, Intent_Standpoints
+from candidates.forms import Intent_StandpointsForm
 from standpoints.models import Standpoints, User_Standpoint
 from users.models import Achievements
 from commontag.views import paginate
@@ -60,24 +62,44 @@ def bills(request, county):
 
 def bill(request, bill_id):
     bill = get_object_or_404(Bills, uid=bill_id)
+    intent = None
     if request.user.is_authenticated():
+        try:
+            intent = Intent.objects.get(user=request.user)
+        except:
+            instance = None
+        else:
+            try:
+                instance = Intent_Standpoints.objects.get(intent=intent, bill_id=bill_id)
+            except:
+                instance = None
+        if intent:
+            form = Intent_StandpointsForm(instance=instance)
         if request.POST:
             with transaction.atomic():
-                if request.POST.get('keyword', '').strip():
-                    standpoint_id = u'bill-%s-%s' % (bill_id, request.POST['keyword'].strip())
-                    standpoint, created = Standpoints.objects.get_or_create(uid=standpoint_id, county=bill.county, title=request.POST['keyword'].strip(), bill_id=bill_id, user=request.user)
-                    if created:
-                        User_Standpoint.objects.create(standpoint_id=standpoint_id, user=request.user)
-                        Standpoints.objects.filter(uid=standpoint_id).update(pro=F('pro') + 1)
-                        tag_create_achievement(request.user)
-                        tag_pro_achievement(standpoint_id)
-                elif request.POST.get('pro'):
-                    User_Standpoint.objects.create(standpoint_id=request.POST['pro'], user=request.user)
-                    Standpoints.objects.filter(uid=request.POST['pro']).update(pro=F('pro') + 1)
-                    tag_pro_achievement(request.POST['pro'])
-                elif request.POST.get('against'):
-                    User_Standpoint.objects.get(standpoint_id=request.POST['against'], user=request.user).delete()
-                    Standpoints.objects.filter(uid=request.POST['against']).update(pro=F('pro') - 1)
+                if request.POST.has_key('intent') and intent:
+                    form = Intent_StandpointsForm(request.POST, instance=instance)
+                    if form.has_changed() and form.is_valid():
+                        intent_sp = form.save(commit=False)
+                        intent_sp.intent = intent
+                        intent_sp.bill = bill
+                        intent_sp.save()
+                else:
+                    if request.POST.get('keyword', '').strip():
+                        standpoint_id = u'bill-%s-%s' % (bill_id, request.POST['keyword'].strip())
+                        standpoint, created = Standpoints.objects.get_or_create(uid=standpoint_id, county=bill.county, title=request.POST['keyword'].strip(), bill_id=bill_id, user=request.user)
+                        if created:
+                            User_Standpoint.objects.create(standpoint_id=standpoint_id, user=request.user)
+                            Standpoints.objects.filter(uid=standpoint_id).update(pro=F('pro') + 1)
+                            tag_create_achievement(request.user)
+                            tag_pro_achievement(standpoint_id)
+                    elif request.POST.get('pro'):
+                        User_Standpoint.objects.create(standpoint_id=request.POST['pro'], user=request.user)
+                        Standpoints.objects.filter(uid=request.POST['pro']).update(pro=F('pro') + 1)
+                        tag_pro_achievement(request.POST['pro'])
+                    elif request.POST.get('against'):
+                        User_Standpoint.objects.get(standpoint_id=request.POST['against'], user=request.user).delete()
+                        Standpoints.objects.filter(uid=request.POST['against']).update(pro=F('pro') - 1)
 
     standpoints_of_bill = Standpoints.objects.filter(bill_id=bill_id)\
                                              .order_by('-pro')
@@ -85,4 +107,4 @@ def bill(request, bill_id):
         standpoints_of_bill = standpoints_of_bill.extra(select={
             'have_voted': "SELECT true FROM standpoints_user_standpoint su WHERE su.standpoint_id = standpoints_standpoints.uid AND su.user_id = %s" % request.user.id,
         },)
-    return render(request, 'bills/bill.html', {'county': bill.county, 'bill': bill, 'standpoints_of_bill': standpoints_of_bill})
+    return render(request, 'bills/bill.html', {'county': bill.county, 'bill': bill, 'standpoints_of_bill': standpoints_of_bill, 'intent': intent, 'form': intent and form})
